server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.0.1
redis==5.2.1
tenacity==9.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
google-auth==2.37.0
email-validator==2.1.1
requests==2.32.3

# Dev / Testing
pytest==8.3.4
pytest-asyncio==0.25.0
ruff==0.8.6
